refactor(clean_up): restructure package and update references
- Merged `lexer` and `parser` package references into a unified structure.
- Updated grammar definitions in `VentoGrammarConfig.bnf` with new package paths.
- Adjusted `plugin.xml` to reflect refactored implementation class references.
- Cleaned up unused imports and improved package organization throughout the codebase.
