+++ src/main/kotlin/org/js/vento/plugin/highlighting/VentoSyntaxHighlighter.kt
@@ -11,8 +11,8 @@ import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.openapi.editor.colors.TextAttributesKey.createTextAttributesKey
 import com.intellij.openapi.fileTypes.SyntaxHighlighterBase
 import com.intellij.psi.tree.IElementType
+import org.js.vento.plugin.VentoTypes
 import org.js.vento.plugin.lexer.VentoLexerAdapter
-import org.js.vento.plugin.lexer.VentoTypes
 
 
 class VentoSyntaxHighlighter : SyntaxHighlighterBase() {
